+++ include/web_video_server/streamers/ros_compressed_streamer.hpp
@@ -30,88 +30,53 @@
 
 #pragma once
 
-extern "C"
-{
-#include <libavcodec/avcodec.h>
-#include <libavformat/avformat.h>
-#include <libavutil/intreadwrite.h>
-#include <libavformat/avio.h>
-#include <libswscale/swscale.h>
-#include <libavutil/opt.h>
-#include <libavutil/mathematics.h>
-#include <libavutil/imgutils.h>
-}
-
-#include <chrono>
 #include <memory>
 #include <string>
 
-#include "image_transport/image_transport.hpp"
-#include "web_video_server/image_streamer.hpp"
-#include "async_web_server_cpp/http_request.hpp"
 #include "async_web_server_cpp/http_connection.hpp"
+#include "async_web_server_cpp/http_request.hpp"
+#include "sensor_msgs/msg/compressed_image.hpp"
+
+#include "web_video_server/multipart_stream.hpp"
+#include "web_video_server/streamers/image_transport_streamer.hpp"
 
 namespace web_video_server
 {
 
-class LibavStreamer : public ImageTransportImageStreamer
+class RosCompressedStreamer : public ImageStreamer
 {
 public:
-  LibavStreamer(
+  RosCompressedStreamer(
     const async_web_server_cpp::HttpRequest & request,
     async_web_server_cpp::HttpConnectionPtr connection,
-    rclcpp::Node::SharedPtr node, const std::string & format_name, const std::string & codec_name,
-    const std::string & content_type);
-
-  ~LibavStreamer();
+    rclcpp::Node::SharedPtr node);
+  ~RosCompressedStreamer();
+  virtual void start();
+  virtual void restreamFrame(std::chrono::duration<double> max_age);
 
 protected:
-  virtual void initializeEncoder();
-  virtual void sendImage(const cv::Mat &, const std::chrono::steady_clock::time_point & time);
-  virtual void initialize(const cv::Mat &);
-  AVFormatContext * format_context_;
-  const AVCodec * codec_;
-  AVCodecContext * codec_context_;
-  AVStream * video_stream_;
-
-  AVDictionary * opt_;   // container format options
+  virtual void sendImage(
+    const sensor_msgs::msg::CompressedImage::ConstSharedPtr msg,
+    const std::chrono::steady_clock::time_point & time);
 
 private:
-  AVFrame * frame_;
-  struct SwsContext * sws_context_;
-  std::mutex encode_mutex_;
-  bool first_image_received_;
-  std::chrono::steady_clock::time_point first_image_time_;
-
-  std::string format_name_;
-  std::string codec_name_;
-  std::string content_type_;
-  int bitrate_;
-  int qmin_;
-  int qmax_;
-  int gop_;
-
-  uint8_t * io_buffer_;  // custom IO buffer
+  void imageCallback(const sensor_msgs::msg::CompressedImage::ConstSharedPtr msg);
+  MultipartStream stream_;
+  rclcpp::Subscription<sensor_msgs::msg::CompressedImage>::SharedPtr image_sub_;
+  std::chrono::steady_clock::time_point last_frame_;
+  sensor_msgs::msg::CompressedImage::ConstSharedPtr last_msg;
+  std::mutex send_mutex_;
+  std::string qos_profile_name_;
 };
 
-class LibavStreamerType : public ImageStreamerType
+class RosCompressedStreamerType : public ImageStreamerType
 {
 public:
-  LibavStreamerType(
-    const std::string & format_name, const std::string & codec_name,
-    const std::string & content_type);
-
   std::shared_ptr<ImageStreamer> create_streamer(
     const async_web_server_cpp::HttpRequest & request,
     async_web_server_cpp::HttpConnectionPtr connection,
     rclcpp::Node::SharedPtr node);
-
   std::string create_viewer(const async_web_server_cpp::HttpRequest & request);
-
-private:
-  const std::string format_name_;
-  const std::string codec_name_;
-  const std::string content_type_;
 };
 
 }  // namespace web_video_server
